Solved day 4 part 1.
I also had a realization that I was repeating code.
Each AOC challenge seems to have a common
pattern in the input. Each line corresponds to an object
or struct of some kind.

I have written a function `read_to_struct` that accepts a
filename and returns a Vec<T> where T must implement
the FromStr trait.

This should signifigantly reduce repitition in the file
reads.