Finally completed day 7.
I had a very hard time with this one.
In the end I wound up going for an arena-allocated tree.

Basically, this means I didn't really make a tree.
I used a vector to store all of my directories,
and I use a usize as the index to reference subdirectories.